Evaluate $PROMPT in __p9__p9k_prepare_prompts
By evaluating $__p9k_unsafe_PROMPT here in __p9k_prepare_prompts we avoid
the evaluation of $PROMPT being interrupted.
For security $PROMPT is never set directly. This way the prompt render is
forced to evaluate the variable and the contents of $__p9k_unsafe_PROMPT
are never executed. The same applies to $RPROMPT.